refactor(core): give the mutation test wrappers a neutral home
