+++ crates/loonfs-core/src/path/write/planner.rs
@@ -198,12 +198,12 @@ fn attribute(error: CoreError, index: usize, operation_count: usize) -> CoreErro
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::checkpoint::tests::ops::{delete_path, put_file_bytes};
     use crate::commit::{CandidateAllocation, InodeAllocator};
     use crate::context::MutationContext;
     use crate::namespace::bootstrap::bootstrap_namespace;
     use crate::path::read::load_current_metadata_view;
     use crate::storage::content::store_bytes_as_content;
+    use crate::test_support::ops::{delete_path, put_file_bytes};
     use loonfs_api::{
         AbsolutePath, CommitId, DeleteDirectoryBehavior, DestinationBehavior, InodeId,
     };
