+++ crates/loonfs-core/src/test_support/content_write.rs
